Add Let's Encrypt setup script.
Let's Encrypt provides free TLS/SSL certificates, which is trusted by
most platforms [ https://letsencrypt.org ].

This setup script, helps with the initial setup, but also adds a cronjob
to check weekly if it needs to be renewed (and does so if needed).

It also bounces lighttp if the cert is replaced.

Nothing OOTB changes, and requires these steps to enable:
cd /system/sdcard/config/ssl/
./letsencrypt_setup.sh

The domain (which must be valid, and reachable via :80 from the
internet) is stored here:
/system/sdcard/config/ssl_domain.conf

Signed-off-by: Dave Walker (Daviey) <email@daviey.com>