Add standard periodic cronjobs (#479)
Previously adding a cronjob meant adding it to the root crontab with
appropriate syntax.  This change adds some standard periodic locations,
where simply dropping in a shell script into a folder means it will be
executed on schedule.  Adding specific jobs remains unaffected by this
change, and upgrades will not be touched.

The periodic jobs are:
  - 15min (every 15 mins)
  - hourly (on the hour)
  - daily (02:00 everyday)
  - weekly (03:00 on Saturday)
  - monthly (05:00 on first of month)

ie, adding a script to:
 - /system/sdcard/config/cron/periodic/15min/foo.sh
will execute every 15 mins.

Signed-off-by: Dave Walker (Daviey) <email@daviey.com>