Fix CGI execution and add cgi stderr logging
Change #478 introduced a regression with the updated busybox where ".",
seems not be part of the default $PATH.

The cgi scripts were sourcing "func.cgi", but this could not be found in
the new default $PATH, and therefore failed to execute. This meant that
the web-ui was unable to make changes and execute commands.

This has been resolved by putting a leading "./" for the script, making
it now ./func.cgi.

In addition, this also adds the lighttpd option server.breakagelog,
which enables stderr logging to /tmp/lighttpd-cgi-stderr.log (so 500
internal errors output is now logged to file (previously just the error
500 was in the webserver logs).

Closes: #491

Signed-off-by: Dave Walker (Daviey) <email@daviey.com>