Update busybox to include flash_eraseall
Previously flash_eraseall worked, because the vendor provided busybox
included the applet flash_eraseall.  However, the system busybox was
recently updated to point to the custom one.

This was missing the applet flash_eraseall, which this change fixes.  In
addition, this change updates the ~root/.busybox_aliases with the alias
(and verified there were no other missing applets).

Closes: #496

Signed-off-by: Dave Walker (Daviey) <email@daviey.com>