Provide the captured image as a parameter to userscripts on motion detection
This patch allows the user scripts to act directly on the captured image
by passing the path to the scripts directly as the second argument.
This way it's possible to work with the image directly, without
employing dirty tricks to find out which one was the latest image in the
stills directory.