Log by default to /tmp so as to avoid SD card corruption
I discovered that the errors I was getting about the SD card similar to #880 were always related to the v4l2rtspserver log.
It seems to me that especially when turning detection tracking due to the high number of writes, writing to the log causes SD card corruption. I suppose that the SD card controller is not built for such a speed.
Since changing logging to log to /tmo which is mounted as tmpfs, I hadn't had any issues.