Change configuration files to have *.dist extenstion
Files with the *.conf extension are skipped during autoupdate. We shuld ensure that no files with
*.user/*.conf are uploaded to the repository.

If someone is up to it, it would be nice to check whether any of the dist files has been updated.
If no such file exists autoupdate should create a local copy without *.dist. If one exists, a diff
should be shown to the user. Or the missing keys should be automagically added to the existing local
copy.