Add Jerusalem timezone to timezones.tsv (#918)
* Add Jerusalem timezone to timezones.tsv

This has been added ata later date. See https://dev.archive.openwrt.org/ticket/11445.html

* Change Jerusalem timezone to change at an earlier offset

uclibc which is used to parse the /etc/TZ file has a problem parsing Israel's DST rule. This is because
the rule employs an offset greater than 24 hours. A mail has been sent to the uclibc mailing list and
an issue has been raised in Dafang-Hacks/Main#30.

As a temporary workaround I changed the rule from 26 to 23, which will cause the clock to change 3
hours earlier when moving to DST.