Optimize tracking detection (#926)
* detectionTracking improvements.

Remove sleep statement when running back to origin. Sleep is alredy
delat with in PTZ script.

Move sanity checks to top, which frees us from nesting caculations.

Send motor commands to background. It shouldnt be rin in the same
thread. This might solve some more time out issues with home
assistant. Add a sleep command at the end, allowing thhe motor to
run before we finish the script.

* Optimized too much

Get script running correctly again

* Fix case of UP and DOWN/LEFT and RIGHT

We can't return 0. As sometimes we have UP and DOWN both on the same side, which means we need to go left/right
Same goes for RIGHT and LEFT.

* Ternary operator doesn't work

So use regular if statements.
Also the sleep time is not long enough, using trial and error discovered 0.5 to be long enough.

* Comments and cleanup