Update axis info in OSD when tracking motion
When tracking motion we should show the updated value of the axis
on the osd.

I discovered along the way that we already have builtin movement commands
in the common functions which deal with all of this.

Along the way I moved the actual movement to be done in the
background, hoping it might alleviate some of the timeouts seen
when polling every 10 seconds for an image like in home assistant.
Unfortunately, this didnt help. Still leaving it in there as it seems
to me to be correct way of doing this (TM) ;-) .