Build cfw-1.2 for Wyze V2 with firmware 4.9.4.28
This provides a new cfw-1.2.bin firmware rebuilt on the official 4.9.4.28
firmware, so that the SD card with the Xiaomi-Dafang-Hacks firmware_mod
can be removed and the camera will boot on the official 4.9.4.28 firmware.

This makes Wyze V2 cameras with the jxf23 cmos sensor work again when
the SD card is removed (https://forums.wyzecam.com/t/wyze-cam-v2-component-change-transition-to-jx-f23-cmos-sensor/6286).
The current cfw-1.1.bin is built on a very old firmware (4.9.1.60) that doesn't
support jxf23.