[Push To Talk] Added functionality to playback sound from the web
Created a simple UI that has a Record and Send button.

Using standard javascript I create a WAV in 8KHZ that works with the audioplay binary.

It is a simple push to talk.

Note:

I find an issue with the audio streaming, it seems that audioplay changes the bitrate of the audio
and i have to restart the rtsp if I want to hear back correctly.