Fix 'rtsp-h264' still set as an autostart entry when 'rtsp' is now the service to use
In 1e2f96612f4ad7852697609669d72532aea2a314 the services 'rtsp-h264' and
'rtsp-mjpeg' were combined into a single 'rtsp' service, however an
entry for 'rtsp-h264' still remained in the autostart folder. This is a
problem for the user because this entry does not appear in the Web UI
which means it cannot be easily removed, and also appears as if an RTSP
server is not running when it is.