Fix update w3.css
The previous PR I submitted was supposed to make the footer object stick to the bottom of the page but it did not work. This was because I missed a comma when adding the relevant changes.

This PR changes one comma in w3.css to a semi-colon, fixing the desired footer behaviour.