This adds a new feature "Dropbox URL" so you can specify the URL to be used for Dropbox uploads. What this allows is running your own server to store snapshots and videos.
There is a new config option "dropbox_url" that defaults to the real Dropbox.

This patch includes a PHP script that you can run on your server to service
uploads from cameras:

   * integration/custom/dafangstor.php

Some documentation on how this works is included as well:

   * integration/custom/webserverstorage.md

NOTE: Dropbox will be discontinuing this interface in I believe September 2021
so after that running your own server may be the only option.